Tokens are required in jobsub_lite infrastructure for submission.  So keep underlying creds library same and enforce infrastructure rule at get_parser parser
