+++ tests/data/auth_methods_args_bad.json
@@ -1,5 +1,9 @@
 [
     {
+        "cmdline_args": "proxy",
+        "bad_auth_method": "proxy"
+    },
+    {
         "cmdline_args": "token,BADSTUFF",
         "bad_auth_method": "BADSTUFF"
     },
